Fixes: The "path" argument must be of type string. Received type undefined.
There's an issue with the way that source code is found ( referreneced in issue #101 ). If you're including solidity files from your node modules, this plugin won't find them. 

This fixes the problem, but I'm pretty sure there is a more eloquent way to solve this. Sorry for the ugly code.